benchmark/04-crewai-research-crew/requirements.txt
crewai>=0.80.0,<0.90.0
requests>=2.31.0
python-dotenv>=1.0.0
